+++ game-aplaguetalerequiem/CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - None
 
+## [1.3.1] - 2025-11-17
+
+- Added installer for folders in game root to help remove extraneous top-level folders.
+
 ## [1.3.0] - 2025-11-14
 
 - Fixed game version detection for Xbox version.