+++ game-darkmessiahofmightandmagic/CHANGELOG.md
@@ -1,13 +1,22 @@
 # Changelog
 
+## [0.2.3] - 2025-11-18
+
+- Added config.cfg file installer.
+- Added buttons to open config.cfg and Saves folder (folder icon on Mods page toolbar).
+- Technical fixes and improvements.
+
 ## [0.2.2]
+
 - Removed 'bin' folder from the game root installer because it can conflict with some Launcher Mods
 
 ## [0.2.1]
+
 - Added installer for additional folders in the game root.
 - Fixed typo in a modtype name
 
 ## [0.2.0]
+
 - Changed URL for browsing to wiltOS Mod Launcher download to ensure user finds the latest version
 - Fixed unhandled exception when downloading wiltOS Mod Launcher
 - Cleaned up the code