+++ game-monsterhunterwilds/CHANGELOG.md
@@ -4,10 +4,11 @@
 
 - None Planned
 
-## [0.2.2] - 2025-11-16
+## [0.2.2] - 2025-11-18
 
+- Added a Root Folder installer to try and filter mods that are not for Fluffy from being re-zipped. Targets any mod with a .exe file and a few dlls for now.
 - Added buttons to open useful files and folder (inside folder icon in Mods page toolbar).
-- Added support for Benchmark demo.
+- Added discovery for Benchmark demo version.
 - Technical fixes and improvements.
 
 ## [0.2.1] - 2025-07-01